import 9.020 (AladinBetaSrc.jar 9.020)

+++ cds/allsky/Context.java
@@ -24,6 +24,7 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -1200,11 +1201,11 @@ public class Context {
 
       long nbTilesPerMin = (deltaNbTile*60000L)/deltaTime;
 
-      String s=statNbTile+"+"+statNbEmptyTile+sNbCells+" tiles + "+statNodeTile+" nodes in "+Util.getTemps(totalTime,true)+" ("
-            +pourcentNbCells+(nbTilesPerMin<=0 ? "": " "+nbTilesPerMin+"tiles/mn EndsIn:"+Util.getTemps(tempsTotalEstime,true))+") "
-            +Util.getTemps(statAvgTime)+"/tile ["+Util.getTemps(statMinTime)+" .. "+Util.getTemps(statMaxTime)+"] "
+      String s=statNbTile+(statNbEmptyTile==0?"":"+"+statNbEmptyTile)+sNbCells+" tiles + "+statNodeTile+" nodes in "+Util.getTemps(totalTime,true)+" ("
+            +pourcentNbCells+(nbTilesPerMin<=0 ? "": " "+nbTilesPerMin+" tiles/mn EndsIn:"+Util.getTemps(tempsTotalEstime,true))+") "
+//            +Util.getTemps(statAvgTime)+"/tile ["+Util.getTemps(statMinTime)+" .. "+Util.getTemps(statMaxTime)+"] "
 //            +Util.getTemps(statNodeAvgTime)+"/node"
-            +(statNbThread==0 ? "":" by "+statNbThreadRunning+"/"+statNbThread+" threads")
+            +(statNbThread==0 ? "":"by "+statNbThreadRunning+"/"+statNbThread+" threads")
             //         +" using "+Util.getUnitDisk(usedMem)
             ;
 
@@ -1266,11 +1267,31 @@ public class Context {
 
       stat(s);
    }
+   
+   // Demande d'affichage des stats (dans le TabJpeg)
+   protected void showRGBStat(int statNbFile, long cTime,int statNbThread,int statNbThreadRunning) {
+      long nbLowCells = getNbLowCells();
 
-   // Demande d'affichage des stats (dans le TabRgb)
-   protected void showRgbStat(int statNbFile, long statSize, long totalTime) {
-      if( statNbFile>0 ) showJpgStat(statNbFile, totalTime, 1, 1);
+      double pourcent = nbLowCells<=0 ? 0 : (double)statNbFile/nbLowCells;
+      long totalTime = (long)( cTime/pourcent );
+      long endsIn = totalTime-cTime;
+      String pourcentNbCells = nbLowCells==-1 ? "" :
+         (Math.round( ( (double)statNbFile/nbLowCells )*1000)/10.)+"%) ";
+
+      String s;
+      if( nbLowCells<=0 ) s = s=statNbFile+" tiles in "+Util.getTemps(cTime,true);
+      else s=statNbFile+"/"+nbLowCells+" tiles in "+Util.getTemps(cTime,true)+" ("
+            +pourcentNbCells+" endsIn:"+Util.getTemps(endsIn,true)
+            +(statNbThread==0 ? "":" by "+statNbThreadRunning+"/"+statNbThread+" threads");
+
+      stat(s);
    }
+
+
+//   // Demande d'affichage des stats (dans le TabRgb)
+//   protected void showRgbStat(int statNbFile, long statSize, long totalTime) {
+//      if( statNbFile>0 ) showJpgStat(statNbFile, totalTime, 1, 1);
+//   }
 
    protected Action action=null;      // Action en cours (voir Action)
    protected double progress=-1;       // Niveau de progression de l'action en cours, -1 si non encore active, =progressMax si terminée
@@ -1635,8 +1656,23 @@ public class Context {
 
    }
 
-   /** Création, ou mise à jour du fichier des Properties associées au survey */
-   protected void writePropertiesFile() throws Exception {
+   /** Création, ou mise à jour des fichiers meta associées au survey
+    */
+   protected void writeMetaFile() throws Exception {
+      writePropertiesFile(null);
+      
+      // On en profite pour écrire le fichier index.html
+      writeIndexHtml();
+
+      // Et metadata.fits
+      writeMetadataFits();
+
+    }
+   
+   /** Création, ou mise à jour du fichier des Properties associées au survey
+    * @param stream null pour l'écrire à l'emplacement prévu par défaut
+    */
+   protected void writePropertiesFile(OutputStream stream) throws Exception {
 
 
       // Ajout de l'IVORN si besoin
@@ -1683,7 +1719,7 @@ public class Context {
       setPropriete(Constante.KEY_HIPS_TILE_WIDTH,CDSHealpix.pow2( getTileOrder())+"");
 
       // L'url
-      setPropriete("#"+Constante.KEY_HIPS_SERVICE_URL,"ex: http://yourHipsServer/"+label+"");
+      setPropriete("#"+Constante.KEY_HIPS_MASTER_URL,"ex: http://yourHipsServer/"+label+"");
       setPropriete(Constante.KEY_HIPS_STATUS,"public master clonableOnce");
       
       // le status du HiPS : par defaut "public master clonableOnce"
@@ -1714,7 +1750,7 @@ public class Context {
       //      setPropriete(Constante.KEY_HIPS_PROCESS_OVERLAY,
       //            isMap() ? "none" : mode==Mode.ADD ? "add" :
       //               fading ? "border_fading" : mixing ? "mean" : "first");
-      setPropriete(Constante.KEY_HIPS_PROCESS_HIERARCHY, "mean");
+//      setPropriete(Constante.KEY_HIPS_PROCESS_HIERARCHY, jpegMethod.toString().toLowerCase());
 
 
       if( cut!=null ) {
@@ -1796,13 +1832,8 @@ public class Context {
          k[i] = keyAddProp.get(i);
          v[i] = valueAddProp.get(i);
       }
-      updateProperties(k,v,true);
+      updateProperties(k,v,true,stream);
 
-      // On en profite pour écrire le fichier index.html
-      writeIndexHtml();
-
-      // Et metadata.fits
-      writeMetadataFits();
    }
 
    /** Ecriture du fichier des propriétés pour le HpxFinder */
@@ -1836,7 +1867,8 @@ public class Context {
       prop.add("#","#____FOR_COMPATIBILITY_WITH_OLD_HIPS_CLIENTS____");
       prop.add(Constante.OLD_OBS_COLLECTION,label);
       prop.add(Constante.OLD_HIPS_FRAME, getFrameCode() );
-      prop.add(Constante.OLD_HIPS_ORDER,getOrder()+"" );
+      prop.add(Constante.OLD_HIPS_ORDER,prop.getProperty(Constante.KEY_HIPS_ORDER) );
+//    prop.add(Constante.OLD_HIPS_ORDER,getOrder()+"" );
       if( minOrder>3 ) prop.add(Constante.OLD_HIPS_ORDER_MIN, minOrder+"");
       prop.add(Constante.KEY_HIPS_TILE_WIDTH,CDSHealpix.pow2( getTileOrder())+"");
 
@@ -1990,9 +2022,13 @@ public class Context {
     * @param key liste des clés à mettre à jour
     * @param value liste des valuers associées
     * @param overwrite si false, ne peut modifier une clé/valeur déjà existante
+    * @param stream null pour écriture à l'endroit par défaut
     * @throws Exception
     */
    protected void updateProperties(String[] key, String[] value,boolean overwrite) throws Exception {
+      updateProperties(key,value,overwrite,null);
+   }
+   protected void updateProperties(String[] key, String[] value,boolean overwrite,OutputStream stream) throws Exception {
 
       waitingPropertieFile();
       try {
@@ -2002,7 +2038,7 @@ public class Context {
          prop = new MyProperties();
          File f = new File( propFile );
          if( f.exists() ) {
-            if( !f.canRead() || !f.canWrite() ) throw new Exception("Propertie file not available ! ["+propFile+"]");
+            if( !f.canRead() ) throw new Exception("Propertie file not available ! ["+propFile+"]");
             FileInputStream in = new FileInputStream(propFile);
             prop.load(in);
             in.close();
@@ -2059,7 +2095,8 @@ public class Context {
          prop.add("#","#____FOR_COMPATIBILITY_WITH_OLD_HIPS_CLIENTS____");
          prop.add(Constante.OLD_OBS_COLLECTION,getLabel());
          prop.add(Constante.OLD_HIPS_FRAME, getFrameCode() );
-         prop.add(Constante.OLD_HIPS_ORDER,getOrder()+"" );
+         prop.add(Constante.OLD_HIPS_ORDER,prop.getProperty(Constante.KEY_HIPS_ORDER) );
+//         prop.add(Constante.OLD_HIPS_ORDER,getOrder()+"" );
          String fmt = getAvailableTileFormats();
          if( fmt.length()>0 ) prop.add(Constante.OLD_HIPS_TILE_FORMAT,fmt);
          if( fmt.indexOf("fits")>=0 && cut!=null ) {
@@ -2073,21 +2110,24 @@ public class Context {
          }
          
          // Remplacement du précédent fichier
-         String tmp = getOutputPath()+Util.FS+Constante.FILE_PROPERTIES+".tmp";
-         File ftmp = new File(tmp);
-         if( ftmp.exists() ) ftmp.delete();
-         File dir = new File( getOutputPath() );
-         if( !dir.exists() && !dir.mkdir() ) throw new Exception("Cannot create output directory");
-         FileOutputStream out = null;
-         try {
-            out = new FileOutputStream(ftmp);
-            prop.store( out, null);
+         if( stream!=null ) prop.store( stream, null);
+         else {
+            String tmp = getOutputPath()+Util.FS+Constante.FILE_PROPERTIES+".tmp";
+            File ftmp = new File(tmp);
+            if( ftmp.exists() ) ftmp.delete();
+            File dir = new File( getOutputPath() );
+            if( !dir.exists() && !dir.mkdir() ) throw new Exception("Cannot create output directory");
+            FileOutputStream out = null;
+            try {
+               out = new FileOutputStream(ftmp);
+               prop.store( out, null);
 
 
-         } finally {  if( out!=null ) out.close(); }
+            } finally {  if( out!=null ) out.close(); }
 
-         if( f.exists() && !f.delete() ) throw new Exception("Propertie file locked ! (cannot delete)");
-         if( !ftmp.renameTo(new File(propFile)) ) throw new Exception("Propertie file locked ! (cannot rename)");
+            if( f.exists() && !f.delete() ) throw new Exception("Propertie file locked ! (cannot delete)");
+            if( !ftmp.renameTo(new File(propFile)) ) throw new Exception("Propertie file locked ! (cannot rename)");
+         }
 
       }
       finally { releasePropertieFile(); }
@@ -2101,7 +2141,7 @@ public class Context {
          prop = new MyProperties();
          File f = new File( propFile );
          if( f.exists() ) {
-            if( !f.canRead() || !f.canWrite() ) throw new Exception("Propertie file not available ! ["+propFile+"]");
+            if( !f.canRead() ) throw new Exception("Propertie file not available ! ["+propFile+"]");
             FileInputStream in = new FileInputStream(propFile);
             prop.load(in);
             in.close();
